rgw: svc_sysobj: split core out
Split core sysobj interface from sysobj service so that cache could be
implemented.
